move output modules into own folder
