+++ src/output/structured.rs
@@ -1,0 +1,2 @@
+/// A table of data
+pub struct Structured;
